# Conflicts:
#	app/src/main/java/com/example/dayout_organizer/adapter/recyclers/MyTripsAdapter.java
#	app/src/main/java/com/example/dayout_organizer/models/trip/TripData.java
#	app/src/main/java/com/example/dayout_organizer/models/trip/TripModel.java
#	app/src/main/java/com/example/dayout_organizer/ui/fragments/profile/ProfileFragment.java

+++ app/src/main/java/com/example/dayout_organizer/models/trip/CustomerTripData.java
@@ -7,7 +7,7 @@ public class CustomerTripData {
     public int customer_id;
     public int trip_id;
     public int checkout;
-    public int rate;
+    public float rate;
     public String created_at;
     public String updated_at;
     public ProfileUser user = new ProfileUser();
